Change Joypad handling, so inputs are cached during V-Int
It is not recommended to read Joypad input multiple times per frame. For
example, the TH counter on a 6-button controller has to time-out before
it resets, allowing 6-button inputs like Mode/X/Y/Z to be read again. By
reading the inputs again before letting the timer expire, 3-button info
will be read instead of 6-button.

It's best to just only read once per frame, and store the result in RAM,
to be accessed freely without risk of bugs.